Fe criar prova crescer

See merge request !247

+++ api/src/main/java/br/com/cwi/crescer/api/controller/responses/ProvaParaCorrecaoResponse.java
@@ -18,7 +18,7 @@ public class ProvaParaCorrecaoResponse {
 
     private LocalDateTime dataCriacao;
 
-    private int tempoDeDuracaoDaProva;
+    private double tempoDeDuracaoDaProva;
 
     private String nomeCandidato;
